+++ jest.config.js
@@ -1,6 +1,6 @@
 // force timezone to UTC to allow tests to work regardless of local timezone
 // generally used by snapshots, but can affect specific tests
-process.env.TZ = "UTC";
+process.env.TZ = 'UTC';
 
 const {
   grafanaESModules,
@@ -9,7 +9,7 @@ const {
 
 module.exports = {
   // Jest configuration provided by Grafana scaffolding
-  ...require("./.config/jest.config"),
+  ...require('./.config/jest.config'),
   collectCoverageFrom: [
     "src/**/*.{js,jsx,ts,tsx}",
     "!**/node_modules/**",